feat: CMD_MENU_... for menu navigation bindings
This is part of my attempt to somehow get some order out of the
keybindings used across different menus, by putting a layer of
abstraction between individual keys and menu actions. This is only a
partial implementation that focuses on navigation: by design there's
quite a bit of falling back on the old process_key, partly because I'm
undecided about how to handle some of the general cases, and also
because there are many, many, menu-specific bindings that don't seem
like they should obviously be implemented in a general way. Some key
missing things:

* Most aspects of this in webtiles -- I'm completely undecided about how
  or even if to handle this. Many of the keybinds covered in this commit
  have full javascript parallel implementations for snappier UI.
* Many more menu special cases
* select/activate actions
* user rebinding. Right now this is completely disabled, because I think
  there's certain things we would not want to allow at all (e.g.
  bindings for mouse buttons, CMD_MENU_EXIT, stuff like this). Of
  course, this is advanced rc stuff, so it call fall under caveat
  emptor?