refactor: make menu quantity selection non-default
This was a default possibility, I assume, from back in the days where
Menus were used almost entirely for items, and managing stacks mattered
for more things (e.g. weight). Nowadays I'm aware of only two menus that
even need this (drop and pickup), and even here this is not an option
that users have much of a reason to use. Swap the meaning of the flag
and the corresponding defaults so that menus have to enable this rather
than disable it; in principle this also simplifies using digits for
other things in some Menus.